\name{summary}
\alias{summary.estLCCR}
\alias{summary.estLCCRcon}
\alias{summary.confLCCR}

\title{Summary of output}
\description{Summary of output from fitting latent class models for capture-recapture data with covariates.}

\usage{
\method{summary}{estLCCR}(object, ...)
\method{summary}{estLCCRcon}(object, ...)
\method{summary}{confLCCR}(object, ...)
}

\arguments{
\item{object}{output from \code{\link{estLCCR}},\code{\link{estLCCRcon}} or \code{\link{confint.estLCCR}}}
\item{...}{further arguments passed to or from other methods}
}

\value{No return value}

\author{Francesco Bartolucci, Antonio Forcina
}
