\name{print}
\alias{print.estLCCR}
\alias{print.estLCCRcon}
\alias{print.confLCCR}

\title{Print the output}
\description{Given the output, it is written in a readable form.}

\usage{
\method{print}{estLCCR}(x, ...)
\method{print}{estLCCRcon}(x, ...)
\method{print}{confLCCR}(x, ...)
}

\arguments{
  \item{x}{output from \code{\link{estLCCR}}, \code{\link{estLCCRcon}} or \code{\link{confint.estLCCR}}}
  \item{...}{further arguments passed to or from other methods}
}

\value{No return value}

\author{Francesco Bartolucci, Antonio Forcina}
